+++ src/KurzSharp.Templates/GraphQlApi/PlaceholderModelMutation.cs
@@ -24,22 +24,112 @@ public class PlaceholderModelMutation
         _placeholderModelService = placeholderModelService;
     }
 
+    [Error(typeof(GraphQlExceptions.AddEntityException))]
+    [UseMutationConvention(InputArgumentName = "input", InputTypeName = nameof(PlaceholderModelDto),
+        PayloadFieldName = "placeholderModel")]
     public async Task<PlaceholderModelDto> AddPlaceholderModel(PlaceholderModelDto input,
         CancellationToken cancellationToken)
     {
-        return await _placeholderModelService.AddPlaceholderModel(input, cancellationToken);
+        try
+        {
+            return await _placeholderModelService.AddPlaceholderModel(input, cancellationToken);
+        }
+        catch (Exception e)
+        {
+            _logger.LogError("Error while adding {@Entity}, {Message}", input, e.Message);
+
+            throw new GraphQlExceptions.AddEntityException(e.Message);
+        }
     }
 
+    [Error(typeof(GraphQlExceptions.AddEntitiesException))]
+    [UseMutationConvention(InputArgumentName = "input", InputTypeName = nameof(IEnumerable<PlaceholderModelDto>),
+        PayloadFieldName = "placeholderModel")]
+    public async Task<IEnumerable<PlaceholderModelDto>> AddPlaceholderModels(IEnumerable<PlaceholderModelDto> input,
+        CancellationToken cancellationToken)
+    {
+        try
+        {
+            return await _placeholderModelService.AddPlaceholderModels(input, cancellationToken);
+        }
+        catch (Exception e)
+        {
+            _logger.LogError("Error while adding PlaceholderModels, {Message}", e.Message);
+
+            throw new GraphQlExceptions.AddEntitiesException(e.Message);
+        }
+    }
+
+    [Error(typeof(GraphQlExceptions.DeleteEntityException))]
+    [UseMutationConvention(InputArgumentName = "input", InputTypeName = nameof(PlaceholderModelDto),
+        PayloadFieldName = "placeholderModel")]
     public async Task<PlaceholderModelDto> DeletePlaceholderModel(PlaceholderModelDto input,
         CancellationToken cancellationToken)
     {
-        return await _placeholderModelService.DeletePlaceholderModel(input, cancellationToken);
+        try
+        {
+            return await _placeholderModelService.DeletePlaceholderModel(input, cancellationToken);
+        }
+        catch (Exception e)
+        {
+            _logger.LogError("Error while deleting {@Entity}, {Message}", input, e.Message);
+
+            throw new GraphQlExceptions.DeleteEntityException(e.Message);
+        }
     }
 
+    [Error(typeof(GraphQlExceptions.DeleteEntitiesException))]
+    [UseMutationConvention(InputArgumentName = "input", InputTypeName = nameof(IEnumerable<PlaceholderModelDto>),
+        PayloadFieldName = "placeholderModel")]
+    public async Task<IEnumerable<PlaceholderModelDto>> DeletePlaceholderModels(IEnumerable<PlaceholderModelDto> input,
+        CancellationToken cancellationToken)
+    {
+        try
+        {
+            return await _placeholderModelService.DeletePlaceholderModels(input, cancellationToken);
+        }
+        catch (Exception e)
+        {
+            _logger.LogError("Error while deleting PlaceholderModels {Message}", e.Message);
+
+            throw new GraphQlExceptions.DeleteEntitiesException(e.Message);
+        }
+    }
+
+    [Error(typeof(GraphQlExceptions.UpdateEntityException))]
+    [UseMutationConvention(InputArgumentName = "input", InputTypeName = nameof(PlaceholderModelDto),
+        PayloadFieldName = "placeholderModel")]
     public async Task<PlaceholderModelDto> UpdatePlaceholderModel(PlaceholderModelDto input,
         CancellationToken cancellationToken)
     {
-        return await _placeholderModelService.UpdatePlaceholderModel(input, cancellationToken);
+        try
+        {
+            return await _placeholderModelService.UpdatePlaceholderModel(input, cancellationToken);
+        }
+        catch (Exception e)
+        {
+            _logger.LogError("Error while update {@Entity}, {Message}", input, e.Message);
+
+            throw new GraphQlExceptions.UpdateEntityException(e.Message);
+        }
+    }
+
+    [Error(typeof(GraphQlExceptions.UpdateEntitiesException))]
+    [UseMutationConvention(InputArgumentName = "input", InputTypeName = nameof(IEnumerable<PlaceholderModelDto>),
+        PayloadFieldName = "placeholderModel")]
+    public async Task<IEnumerable<PlaceholderModelDto>> UpdatePlaceholderModels(IEnumerable<PlaceholderModelDto> input,
+        CancellationToken cancellationToken)
+    {
+        try
+        {
+            return await _placeholderModelService.UpdatePlaceholderModels(input, cancellationToken);
+        }
+        catch (Exception e)
+        {
+            _logger.LogError("Error while update PlaceholderModels {Message}", e.Message);
+
+            throw new GraphQlExceptions.UpdateEntitiesException(e.Message);
+        }
     }
 #endif
 }