+++ src/KurzSharp.Templates/GrpcApi/IPlaceholderModelGrpcService.cs
@@ -1,3 +1,4 @@
+#nullable enable
 using KurzSharp.Templates.Models;
 #if NET8_0_OR_GREATER
 using System.ServiceModel;
@@ -13,15 +14,27 @@ public interface IPlaceholderModelGrpcService
 {
 #if NET8_0_OR_GREATER
     [OperationContract]
+    Task<PlaceholderModelDto?> GetPlaceholderModel(IdRequest id, CancellationToken cancellationToken);
+
+    [OperationContract]
     Task<IEnumerable<PlaceholderModelDto>> GetPlaceholderModels(CancellationToken cancellationToken);
 
     [OperationContract]
     Task<PlaceholderModelDto> AddPlaceholderModel(PlaceholderModelDto request, CancellationToken cancellationToken);
 
     [OperationContract]
+    Task<IEnumerable<PlaceholderModelDto>> AddPlaceholderModels(IEnumerable<PlaceholderModelDto> requests, CancellationToken cancellationToken);
+
+    [OperationContract]
     Task<PlaceholderModelDto> DeletePlaceholderModel(PlaceholderModelDto request, CancellationToken cancellationToken);
 
     [OperationContract]
+    Task<IEnumerable<PlaceholderModelDto>> DeletePlaceholderModels(IEnumerable<PlaceholderModelDto> requests, CancellationToken cancellationToken);
+
+    [OperationContract]
     Task<PlaceholderModelDto> UpdatePlaceholderModel(PlaceholderModelDto request, CancellationToken cancellationToken);
+
+    [OperationContract]
+    Task<IEnumerable<PlaceholderModelDto>> UpdatePlaceholderModels(IEnumerable<PlaceholderModelDto> requests, CancellationToken cancellationToken);
 #endif
 }