+++ src/KurzSharp.Templates/GrpcApi/PlaceholderModelGrpcService.cs
@@ -1,3 +1,4 @@
+#nullable enable
 #if NET8_0_OR_GREATER
 using ProtoBuf.Grpc.Reflection;
 using ProtoBuf.Meta;
@@ -27,13 +28,13 @@ public partial class PlaceholderModelGrpcService : ControllerBase, IPlaceholderM
         _placeholderModelService = placeholderModelService;
     }
 
-    [HttpGet]
-    public async Task<IEnumerable<PlaceholderModelDto>> GetPlaceholderModels(CancellationToken cancellationToken)
-    {
-        var result = await _placeholderModelService.GetPlaceholderModels(cancellationToken);
+    [HttpGet("{id}")]
+    public async Task<PlaceholderModelDto?> GetPlaceholderModel(IdRequest request, CancellationToken cancellationToken) =>
+        await _placeholderModelService.GetPlaceholderModel(request.Id, cancellationToken);
 
-        return result;
-    }
+    [HttpGet($"/{nameof(PlaceholderModel)}Grpc/bulk")]
+    public async Task<IEnumerable<PlaceholderModelDto>> GetPlaceholderModels(CancellationToken cancellationToken) =>
+        await _placeholderModelService.GetPlaceholderModels(cancellationToken);
 
     [HttpGet("proto")]
     public string GetPlaceholderModelProto(CancellationToken cancellationToken)
@@ -50,7 +51,36 @@ public partial class PlaceholderModelGrpcService : ControllerBase, IPlaceholderM
     public async Task<PlaceholderModelDto> AddPlaceholderModel(PlaceholderModelDto request,
         CancellationToken cancellationToken)
     {
-        return await _placeholderModelService.AddPlaceholderModel(request, cancellationToken);
+        try
+        {
+            return await _placeholderModelService.AddPlaceholderModel(request, cancellationToken);
+        }
+        catch (Exception e)
+        {
+            _logger.LogError("Error while adding {@Entity}, {Message}", request, e.Message);
+
+            throw new RpcException(new Status(Grpc.Core.StatusCode.Internal, $"Error while trying to add {request}"),
+                $"Error while trying to add {request}");
+        }
+    }
+
+    [HttpPost($"/{nameof(PlaceholderModel)}Grpc/bulk")]
+    public async Task<IEnumerable<PlaceholderModelDto>> AddPlaceholderModels(IEnumerable<PlaceholderModelDto> requests,
+        CancellationToken cancellationToken)
+    {
+        try
+        {
+            return await _placeholderModelService.AddPlaceholderModels(requests, cancellationToken);
+        }
+        catch (Exception e)
+        {
+            _logger.LogError("Error while adding PlaceholderModels, {Message}", e.Message);
+
+            throw new RpcException(
+                new Status(Grpc.Core.StatusCode.Internal,
+                    $"Error while trying to add PlaceholderModels, {e.Message}"),
+                $"Error while trying to to add PlaceholderModels, {e.Message}");
+        }
     }
 
     [HttpDelete]
@@ -70,11 +100,61 @@ public partial class PlaceholderModelGrpcService : ControllerBase, IPlaceholderM
         }
     }
 
+    [HttpDelete($"/{nameof(PlaceholderModel)}Grpc/bulk")]
+    public async Task<IEnumerable<PlaceholderModelDto>> DeletePlaceholderModels(
+        IEnumerable<PlaceholderModelDto> requests, CancellationToken cancellationToken)
+    {
+        try
+        {
+            return await _placeholderModelService.DeletePlaceholderModels(requests, cancellationToken);
+        }
+        catch (Exception e)
+        {
+            _logger.LogError("Error while deleting PlaceholderModels {Message}", e.Message);
+
+            throw new RpcException(
+                new Status(Grpc.Core.StatusCode.Internal,
+                    $"Error while trying to delete PlaceholderModels {e.Message}"),
+                $"Error while trying to delete PlaceholderModels {e.Message}");
+        }
+    }
+
     [HttpPut]
     public async Task<PlaceholderModelDto> UpdatePlaceholderModel(PlaceholderModelDto request,
         CancellationToken cancellationToken)
     {
-        return await _placeholderModelService.UpdatePlaceholderModel(request, cancellationToken);
+        try
+        {
+            return await _placeholderModelService.UpdatePlaceholderModel(request, cancellationToken);
+        }
+        catch (Exception e)
+        {
+            _logger.LogError("Error while update {@Entity}, {Message}", request, e.Message);
+
+            throw new RpcException(
+                new Status(Grpc.Core.StatusCode.Internal,
+                    $"Error while trying to update PlaceholderModels {e.Message}"),
+                $"Error while trying to update PlaceholderModels {e.Message}");
+        }
+    }
+
+    [HttpPut($"/{nameof(PlaceholderModel)}Grpc/bulk")]
+    public async Task<IEnumerable<PlaceholderModelDto>> UpdatePlaceholderModels(
+        IEnumerable<PlaceholderModelDto> requests, CancellationToken cancellationToken)
+    {
+        try
+        {
+            return await _placeholderModelService.UpdatePlaceholderModels(requests, cancellationToken);
+        }
+        catch (Exception e)
+        {
+            _logger.LogError("Error while update PlaceholderModels {Message}", e.Message);
+
+            throw new RpcException(
+                new Status(Grpc.Core.StatusCode.Internal,
+                    $"Error while trying to update PlaceholderModels {e.Message}"),
+                $"Error while trying to update PlaceholderModels {e.Message}");
+        }
     }
 }
 #endif