+++ src/KurzSharp/Utils/ModelSourceInfo.cs
@@ -3,18 +3,32 @@ using Microsoft.CodeAnalysis.CSharp.Syntax;
 
 namespace KurzSharp.Utils;
 
+public class ModelProperty(
+    string name,
+    string type,
+    string accessModifier,
+    IReadOnlyCollection<AttributeData> attributes)
+{
+    public string Name { get; } = name;
+    public string Type { get; } = type;
+    public string AccessModifier { get; } = accessModifier;
+    public IReadOnlyCollection<AttributeData> Attributes { get; } = attributes;
+}
+
 public class ModelSourceInfo(
     string typeName,
     string typeNamespace,
     INamedTypeSymbol namedTypeSymbol,
-    IList<AttributeSyntax> attributeSyntaxes) : IEquatable<ModelSourceInfo>
+    IReadOnlyCollection<AttributeSyntax> attributeSyntaxes,
+    IReadOnlyCollection<ModelProperty> properties)
 {
     public string TypeName { get; } = typeName;
     public string TypeNamespace { get; } = typeNamespace;
     public INamedTypeSymbol NamedTypeSymbol { get; } = namedTypeSymbol;
     public IReadOnlyCollection<ApiKind> ApiKinds { get; } = ParseApiKind(attributeSyntaxes);
+    public IReadOnlyCollection<ModelProperty> Properties { get; } = properties;
 
-    private static IReadOnlyCollection<ApiKind> ParseApiKind(IList<AttributeSyntax> attributeSyntaxes) =>
+    private static List<ApiKind> ParseApiKind(IReadOnlyCollection<AttributeSyntax> attributeSyntaxes) =>
         attributeSyntaxes.Select(attributeSyntax =>
         {
             var attributeName = attributeSyntax.Name + "Attribute";
@@ -27,36 +41,4 @@ public class ModelSourceInfo(
                 _ => throw new ArgumentException($"Unknown ApiKind: {attributeSyntax.Name.ToString()}")
             };
         }).ToList();
-
-    public bool Equals(ModelSourceInfo? other)
-    {
-        if (ReferenceEquals(null, other)) return false;
-        if (ReferenceEquals(this, other)) return true;
-
-        return TypeName == other.TypeName && TypeNamespace == other.TypeNamespace &&
-               NamedTypeSymbol.Name == other.NamedTypeSymbol.Name &&
-               ApiKinds.OrderBy(a => a).SequenceEqual(other.ApiKinds.OrderBy(a => a));
-    }
-
-    public override bool Equals(object? obj)
-    {
-        if (ReferenceEquals(null, obj)) return false;
-        if (ReferenceEquals(this, obj)) return true;
-
-        return obj.GetType() == GetType() && Equals((ModelSourceInfo)obj);
-    }
-
-    public override int GetHashCode()
-    {
-        unchecked
-        {
-            var hashCode = TypeName.GetHashCode();
-
-            hashCode = (hashCode * 397) ^ TypeNamespace.GetHashCode();
-            hashCode = (hashCode * 397) ^ NamedTypeSymbol.Name.GetHashCode();
-            hashCode = (hashCode * 397) ^ string.Join(string.Empty, ApiKinds.OrderBy(a => a)).GetHashCode();
-
-            return hashCode;
-        }
-    }
 }